requirements.txt
click>=8.0.0
pyyaml>=6.0
markdown>=3.4.0
jinja2>=3.1.0
anthropic>=0.18.0
beautifulsoup4>=4.12.0
pillow>=10.0.0
watchdog>=3.0.0
requests>=2.31.0
boto3>=1.28.0

